添加了UDP网络系统，网络系统移动到线程client_io

+++ src/com/sunrisestudio/util/SystemInfo.java
@@ -31,6 +31,6 @@ public class SystemInfo {
 
     public static String getUsage(){
         Runtime runtime=Runtime.getRuntime();
-        return 100-(runtime.freeMemory()/ runtime.totalMemory())*100+"%";
+        return (int)(100-(runtime.freeMemory()/ (float)runtime.totalMemory())*100)+"%";
     }
 }
